[refactor] - 방 생성 과정에서 버튼 없애기
