+++ src/components/room/MeetingRoomForm.tsx
@@ -1,35 +1,47 @@
 'use client';
-import TagList from './MeetingRoomFeatureTags';
-import { useRoomStore } from '(@/store/roomStore)';
-import { Button, Modal, ModalBody, ModalContent, ModalFooter, ModalHeader, useDisclosure } from '@nextui-org/react';
-import { useState } from 'react';
 import { useAddRoom } from '(@/hooks/useMutation/useMeetingMutation)';
-import { userStore } from '(@/store/userStore)';
+import { useRoomStore } from '(@/store/roomStore)';
+import { favoriteOptions } from '(@/utils/FavoriteData)';
+import {
+  Button,
+  Chip,
+  Modal,
+  ModalBody,
+  ModalContent,
+  ModalFooter,
+  ModalHeader,
+  Select,
+  SelectItem,
+  useDisclosure
+} from '@nextui-org/react';
 import { useRouter } from 'next/navigation';
-
-import type { NextMeetingRoomType } from '(@/types/roomTypes)';
-import RegionSelection from './RegionSelection';
+import { useState } from 'react';
 import MemberNumberSelection from './MemberNumberSelection';
+import RegionSelection from './RegionSelection';
+
+import { useGetUserDataQuery } from '(@/hooks/useQueries/useUserQuery)';
+import type { MeetingRoomType } from '(@/types/roomTypes)';
 
 function MeetingRoomForm() {
   const router = useRouter();
-  const { tags, resetTags } = useRoomStore();
-  const { isOpen, onOpen, onOpenChange } = useDisclosure();
-  const { user } = userStore((state) => state);
-  const [title, setTitle] = useState('');
   const [location, setLocation] = useState('');
+  const { isOpen, onOpen, onOpenChange } = useDisclosure();
   const { memberNumber, resetMemberNumber, roomRegion, resetRoomRegion } = useRoomStore((state) => state);
+  const { data: user, isPending, isError, error } = useGetUserDataQuery();
+  const [title, setTitle] = useState('');
+  const [selected, setSelected] = useState<Set<string>>(new Set([]));
+
   const cancelMakingMeetingRoom = () => {
     setTitle('');
     setLocation('');
     resetMemberNumber();
     resetRoomRegion();
-    resetTags();
+    setSelected(new Set([]));
   };
-
+  const favoriteArray = Array.from(selected);
   const user_id = user?.user_id!;
-  const nextMeetingRoom: NextMeetingRoomType = {
-    feature: tags,
+  const nextMeetingRoom: MeetingRoomType = {
+    feature: favoriteArray,
     leader_id: String(user_id),
     location,
     region: String(roomRegion),
@@ -41,13 +53,29 @@ function MeetingRoomForm() {
 
   const addMeetingRoom = async (e: any) => {
     e.preventDefault();
-    if (!title || !tags || !location || memberNumber === '인원수' || !roomRegion) {
+    if (!title || !selected || !location || memberNumber === '인원수' || !roomRegion) {
       return alert('모든 항목은 필수입니다.');
     }
-    const newRoom = await addRoomMutation.mutateAsync();
+    const data = await addRoomMutation.mutateAsync();
     alert('모임이 생성되었습니다.');
-    resetTags();
-    router.push(`/meetingRoom/${newRoom}`);
+    setSelected(new Set([]));
+    resetMemberNumber();
+    resetRoomRegion();
+    router.push(`/meetingRoom/${data}`);
+  };
+
+  const handleSelect = (value: string[]) => {
+    if (selected.size > 5) {
+      alert('최대 5개까지 선택 가능합니다.');
+      return;
+    }
+    setSelected(new Set(value));
+  };
+
+  const handleDelete = (value: string) => {
+    const newSelected = new Set(selected);
+    newSelected.delete(value);
+    setSelected(newSelected);
   };
 
   return (
@@ -87,7 +115,38 @@ function MeetingRoomForm() {
               <form onSubmit={(e) => addMeetingRoom(e)}>
                 <ModalBody>
                   <input value={title} onChange={(e) => setTitle(e.target.value)} placeholder="title" maxLength={15} />
-                  <TagList />
+                  <div className="flex w-full max-w-xs flex-col gap-2">
+                    <label>방의 컨셉을 골라주세요!</label>
+                    <div className="flex whitespace-nowrap">
+                      <Select
+                        label="방의 특성(최대 5개)"
+                        selectionMode="multiple"
+                        variant="bordered"
+                        selectedKeys={selected}
+                        className="max-w-xs"
+                        aria-label="방의 특성"
+                        onSelectionChange={(value) => handleSelect(value as string[])}
+                      >
+                        {favoriteOptions.map((option) => (
+                          <SelectItem key={option.value} value={option.value}>
+                            {option.value}
+                          </SelectItem>
+                        ))}
+                      </Select>
+                    </div>
+                    <div className="flex flex-wrap gap-2">
+                      {Array.from(selected).map((value) => (
+                        <Chip
+                          key={value}
+                          color="default"
+                          style={{ backgroundColor: favoriteOptions.find((option) => option.value === value)?.color }}
+                          onClose={() => handleDelete(value)}
+                        >
+                          {value}
+                        </Chip>
+                      ))}
+                    </div>
+                  </div>
                   <div className="flex flex-row gap-4">
                     <MemberNumberSelection text={'member'} />
                     <div className="w-20">
